Breusch-Godfrey Test

bgtest performs the Breusch-Godfrey test for higher-order serial correlation.

breusch_godfrey(x, ...)

Value

A list with class "bgtest" inheriting from "htest" containing the following components:

statistic

the value of the test statistic.

p.value

the p-value of the test.

parameter

degrees of freedom.

method

a character string indicating what type of test was performed.

data.name

a character string giving the name(s) of the data.

coefficients

coefficient estimates from the auxiliary regression.

vcov

corresponding covariance matrix estimate.

Details

Under \(H_0\) the test statistic is asymptotically Chi-squared with degrees of freedom as given in parameter. If type is set to "F" the function returns a finite sample version of the test statistic, employing an \(F\) distribution with degrees of freedom as given in parameter.

By default, the starting values for the lagged residuals in the auxiliary regression are chosen to be 0 (as in Godfrey 1978) but could also be set to NA to omit them.

bgtest also returns the coefficients and estimated covariance matrix from the auxiliary regression that includes the lagged residuals. Hence, coeftest can be used to inspect the results. (Note, however, that standard theory does not always apply to the standard errors and t-statistics in this regression.)

References

Breusch, T.S. (1978): Testing for Autocorrelation in Dynamic Linear Models, Australian Economic Papers, 17, 334-355.

Godfrey, L.G. (1978): Testing Against General Autoregressive and Moving Average Error Models when the Regressors Include Lagged Dependent Variables', Econometrica, 46, 1293-1301.

Wooldridge, J.M. (2013): Introductory Econometrics: A Modern Approach, 5th edition, South-Western College.

See also

Author

David Mitchell <david.mitchell@dotars.gov.au>, Achim Zeileis

Examples

## Generate a stationary and an AR(1) series
     x <- rep(c(1, -1), 50)

     y1 <- 1 + x + rnorm(100)

     ## Perform Breusch-Godfrey test for first-order serial correlation:
     bgtest(y1 ~ x)

#> Error in bgtest(y1 ~ x): could not find function "bgtest"

     ## or for fourth-order serial correlation
     bgtest(y1 ~ x, order = 4)

#> Error in bgtest(y1 ~ x, order = 4): could not find function "bgtest"

     ## Compare with Durbin-Watson test results:
     dwtest(y1 ~ x)

#> Error in dwtest(y1 ~ x): could not find function "dwtest"

     y2 <- filter(y1, 0.5, method = "recursive")

#> Error in UseMethod("filter"): no applicable method for 'filter' applied to an object of class "c('double', 'numeric')"

     bgtest(y2 ~ x)

#> Error in bgtest(y2 ~ x): could not find function "bgtest"

     bg4 <- bgtest(y2 ~ x, order = 4)

#> Error in bgtest(y2 ~ x, order = 4): could not find function "bgtest"

     bg4

#> Error in eval(expr, envir, enclos): object 'bg4' not found

     coeftest(bg4)

#> Error in coeftest(bg4): could not find function "coeftest"