fable: Forecasting Models for Tidy Time Series

Provides a collection of commonly used univariate and multivariate time series forecasting models including automatically selected exponential smoothing (ETS) and autoregressive integrated moving average (ARIMA) models. These models work within the 'fable' framework provided by the 'fabletools' package, which provides the tools to evaluate, visualise, and combine models in a workflow consistent with the tidyverse.

See also

Author

Maintainer: Mitchell O'Hara-Wild mail@mitchelloharawild.com

Authors:

Rob Hyndman

Earo Wang

Other contributors:

Gabriel Caceres (NNETAR implementation) [contributor]

Christoph Bergmeir (ORCID) [contributor]

Tim-Gunnar Hensel [contributor]

Timothy Hyndman [contributor]