Generate new data from a fable model

Simulates future paths from a dataset using a fitted model. Innovations are sampled by the model's assumed error distribution. If bootstrap is TRUE, innovations will be sampled from the model's residuals. If new_data contains the .innov column, those values will be treated as innovations.

# S3 method for NNETAR
generate(x, new_data, specials = NULL, bootstrap = FALSE, ...)

Arguments

x
A fitted model.

new_data
A tsibble containing future information used to forecast.

specials
(passed by fabletools::forecast.mdl_df()).

bootstrap
If TRUE, then forecast distributions are computed using simulation with resampled errors.

...
Additional arguments for forecast model methods.

See also

Examples

as_tsibble(airmiles) %>%
  model(nn = NNETAR(box_cox(value, 0.15))) %>%
  generate()

#> # A tsibble: 2 x 4 [1Y]
#> # Key:       .model, .rep [1]
#>   .model index .rep    .sim
#>   <chr>  <dbl> <chr>  <dbl>
#> 1 nn      1961 1     33926.
#> 2 nn      1962 1     34189.