Generate new data from a fable model

Simulates future paths from a dataset using a fitted model. Innovations are sampled by the model's assumed error distribution. If bootstrap is TRUE, innovations will be sampled from the model's residuals. If new_data contains the .innov column, those values will be treated as innovations.

# S3 method for RW
generate(x, new_data, bootstrap = FALSE, ...)

Arguments

x

A fitted model.

new_data

A tsibble containing future information used to forecast.

bootstrap

If TRUE, then forecast distributions are computed using simulation with resampled errors.

...

Additional arguments for forecast model methods.

See also

Examples

as_tsibble(Nile) %>%
  model(NAIVE(value)) %>%
  generate()
#> # A tsibble: 2 x 4 [1Y]
#> # Key:       .model, .rep [1]
#>   .model       index .rep   .sim
#>   <chr>        <dbl> <chr> <dbl>
#> 1 NAIVE(value)  1971 1      913.
#> 2 NAIVE(value)  1972 1     1070.

library(tsibbledata)
aus_production %>%
  model(snaive = SNAIVE(Beer ~ lag("year"))) %>%
  generate()
#> # A tsibble: 8 x 4 [1Q]
#> # Key:       .model, .rep [1]
#>   .model Quarter .rep   .sim
#>   <chr>    <qtr> <chr> <dbl>
#> 1 snaive 2010 Q3 1      424.
#> 2 snaive 2010 Q4 1      528.
#> 3 snaive 2011 Q1 1      404.
#> 4 snaive 2011 Q2 1      391.
#> 5 snaive 2011 Q3 1      430.
#> 6 snaive 2011 Q4 1      535.
#> 7 snaive 2012 Q1 1      410.
#> 8 snaive 2012 Q2 1      429.