Generate new data from a fable model

Simulates future paths from a dataset using a fitted model. Innovations are sampled by the model's assumed error distribution. If bootstrap is TRUE, innovations will be sampled from the model's residuals. If new_data contains the .innov column, those values will be treated as innovations.

# S3 method for model_mean
generate(x, new_data, bootstrap = FALSE, ...)

Arguments

x
A fitted model.

new_data
A tsibble containing future information used to forecast.

bootstrap
If TRUE, then forecast distributions are computed using simulation with resampled errors.

...
Additional arguments for forecast model methods.

See also

Examples

library(tsibbledata)
vic_elec %>%
  model(avg = MEAN(Demand)) %>%
  generate()

#> # A tsibble: 4 x 4 [30m] <UTC>
#> # Key:       .model, .rep [1]
#>   .model  .rep Time                 .sim
#>   <chr>  <int> <dttm>              <dbl>
#> 1 avg        1 2015-01-01 00:00:00 3080.
#> 2 avg        1 2015-01-01 00:30:00 4029.
#> 3 avg        1 2015-01-01 01:00:00 4383.
#> 4 avg        1 2015-01-01 01:30:00 4769.