Glance a average method model

Construct a single row summary of the average method model.

# S3 method for model_mean
glance(x, ...)

Arguments

x
model or other R object to convert to single-row data frame

...
other arguments passed to methods

Value

A one row tibble summarising the model's fit.

Details

Contains the variance of residuals (sigma2).

Examples

library(tsibbledata)
vic_elec %>%
  model(avg = MEAN(Demand)) %>%
  glance()

#> # A tibble: 1 x 2
#>   .model  sigma2
#>   <chr>    <dbl>
#> 1 avg    764354.