Refit a TSLM

Applies a fitted TSLM to a new dataset.

# S3 method for TSLM
refit(object, new_data, specials = NULL, reestimate = FALSE, ...)

Arguments

object
The time series model used to produce the forecasts

new_data
A tsibble containing future information used to forecast.

specials
(passed by fabletools::forecast.mdl_df()).

reestimate
If TRUE, the coefficients for the fitted model will be re-estimated to suit the new data.

...
Additional arguments for forecast model methods.

Examples

lung_deaths_male <- as_tsibble(mdeaths)
lung_deaths_female <- as_tsibble(fdeaths)

fit <- lung_deaths_male %>%
  model(TSLM(value ~ trend() + season()))

report(fit)

#> Series: value 
#> Model: TSLM 
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -422.6976  -58.6048   -0.1143   64.2905  644.0310 
#> 
#> Coefficients:
#>                 Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)    1968.9500    79.6967  24.706  < 2e-16 ***
#> trend()          -4.1060     0.9662  -4.250 7.72e-05 ***
#> season()year1   288.1679    97.5843   2.953  0.00451 ** 
#> season()year2   243.7738    97.4838   2.501  0.01519 *  
#> season()year3   137.0464    97.3927   1.407  0.16463    
#> season()year4  -172.0143    97.3112  -1.768  0.08229 .  
#> season()year5  -511.0750    97.2393  -5.256 2.13e-06 ***
#> season()year6  -634.3024    97.1768  -6.527 1.70e-08 ***
#> season()year7  -680.3631    97.1240  -7.005 2.65e-09 ***
#> season()year8  -775.2571    97.0807  -7.986 5.79e-11 ***
#> season()year9  -787.1512    97.0471  -8.111 3.55e-11 ***
#> season()year10 -563.5452    97.0230  -5.808 2.68e-07 ***
#> season()year11 -422.9393    97.0086  -4.360 5.29e-05 ***
#> season()year12        NA         NA      NA       NA    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 168 on 59 degrees of freedom
#> Multiple R-squared: 0.875,	Adjusted R-squared: 0.8495
#> F-statistic: 34.41 on 12 and 59 DF, p-value: < 2.22e-16

fit %>%
  refit(lung_deaths_female) %>%
  report()

#> Series: value 
#> Model: TSLM 
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -1515.5 -1123.7  -897.9  -748.7  -505.5 
#> 
#> Coefficients:
#>                Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)    1968.950    506.705   3.886 0.000261 ***
#> trend()          -4.106      6.143  -0.668 0.506480    
#> season()year1   288.168    620.432   0.464 0.644025    
#> season()year2   243.774    619.793   0.393 0.695505    
#> season()year3   137.046    619.214   0.221 0.825605    
#> season()year4  -172.014    618.696  -0.278 0.781964    
#> season()year5  -511.075    618.238  -0.827 0.411759    
#> season()year6  -634.302    617.842  -1.027 0.308780    
#> season()year7  -680.363    617.506  -1.102 0.275026    
#> season()year8  -775.257    617.231  -1.256 0.214058    
#> season()year9  -787.151    617.017  -1.276 0.207048    
#> season()year10 -563.545    616.864  -0.914 0.364664    
#> season()year11 -422.939    616.772  -0.686 0.495569    
#> season()year12       NA         NA      NA       NA    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 1068 on 59 degrees of freedom
#> Multiple R-squared: 0.1476,	Adjusted R-squared: -0.0258
#> F-statistic: 0.8512 on 12 and 59 DF, p-value: 0.59898