Tidy a fable model

Returns the coefficients from the model in a tibble format.

# S3 method for ETS
tidy(x, ...)

Arguments

x
An object to be converted into a tidy tibble::tibble().

...
Additional arguments to tidying method.

Value

The model's coefficients in a tibble.

Examples

as_tsibble(USAccDeaths) %>%
  model(ets = ETS(log(value) ~ season("A"))) %>%
  tidy()

#> # A tibble: 17 x 3
#>    .model term    estimate
#>    <chr>  <chr>      <dbl>
#>  1 ets    alpha   0.533   
#>  2 ets    beta    0.0196  
#>  3 ets    gamma   0.000100
#>  4 ets    l[0]    9.20    
#>  5 ets    b[0]   -0.00305 
#>  6 ets    s[0]    0.00184 
#>  7 ets    s[-1]  -0.0292  
#>  8 ets    s[-2]   0.0302  
#>  9 ets    s[-3]  -0.00355 
#> 10 ets    s[-4]   0.110   
#> 11 ets    s[-5]   0.180   
#> 12 ets    s[-6]   0.0930  
#> 13 ets    s[-7]   0.0418  
#> 14 ets    s[-8]  -0.0569  
#> 15 ets    s[-9]  -0.0856  
#> 16 ets    s[-10] -0.190   
#> 17 ets    s[-11] -0.0908