Tidy a fable model

Returns the coefficients from the model in a tibble format.

# S3 method for ARIMA
tidy(x, ...)

Arguments

x

An object to be converted into a tidy tibble::tibble().

...

Additional arguments to tidying method.

Value

The model's coefficients in a tibble.

Examples

USAccDeaths %>%
  as_tsibble() %>%
  model(arima = ARIMA(log(value) ~ pdq(0, 1, 1) + PDQ(0, 1, 1))) %>%
  tidy()
#> # A tibble: 2 × 6
#>   .model term  estimate std.error statistic  p.value
#>   <chr>  <chr>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 arima  ma1     -0.471     0.123     -3.83 0.000311
#> 2 arima  sma1    -0.593     0.193     -3.07 0.00327