Tidy a fable model

Returns the coefficients from the model in a tibble format.

# S3 method for VAR
tidy(x)

Arguments

x
An object to be converted into a tidy tibble::tibble().

Value

The model's coefficients in a tibble.

Examples

lung_deaths <- cbind(mdeaths, fdeaths) %>%
  as_tsibble(pivot_longer = FALSE)

lung_deaths %>%
  model(VAR(vars(log(mdeaths), fdeaths) ~ AR(3))) %>%
  tidy()

#> # A tibble: 14 x 7
#>    .model            term       .response   estimate std.error statistic p.value
#>    <chr>             <chr>      <chr>          <dbl>     <dbl>     <dbl>   <dbl>
#>  1 VAR(vars(log(mde… lag(log(m… log(mdeat…   1.04e+0   2.64e-1    3.92   2.23e-4
#>  2 VAR(vars(log(mde… lag(fdeat… log(mdeat…   1.06e-5   3.81e-4    0.0277 9.78e-1
#>  3 VAR(vars(log(mde… lag(log(m… log(mdeat…   3.05e-1   2.97e-1    1.03   3.08e-1
#>  4 VAR(vars(log(mde… lag(fdeat… log(mdeat…  -7.74e-4   3.87e-4   -2.00   4.99e-2
#>  5 VAR(vars(log(mde… lag(log(m… log(mdeat…  -3.00e-1   2.57e-1   -1.17   2.46e-1
#>  6 VAR(vars(log(mde… lag(fdeat… log(mdeat…   5.31e-5   3.75e-4    0.142  8.88e-1
#>  7 VAR(vars(log(mde… constant   log(mdeat…   9.50e-2   2.07e+0    0.0459 9.64e-1
#>  8 VAR(vars(log(mde… lag(log(m… fdeaths      4.49e+2   1.74e+2    2.58   1.24e-2
#>  9 VAR(vars(log(mde… lag(fdeat… fdeaths      2.89e-1   2.51e-1    1.15   2.55e-1
#> 10 VAR(vars(log(mde… lag(log(m… fdeaths      2.06e+2   1.96e+2    1.05   2.97e-1
#> 11 VAR(vars(log(mde… lag(fdeat… fdeaths     -4.33e-1   2.55e-1   -1.70   9.46e-2
#> 12 VAR(vars(log(mde… lag(log(m… fdeaths     -2.34e+2   1.69e+2   -1.38   1.71e-1
#> 13 VAR(vars(log(mde… lag(fdeat… fdeaths      2.51e-2   2.47e-1    0.102  9.19e-1
#> 14 VAR(vars(log(mde… constant   fdeaths     -2.43e+3   1.37e+3   -1.78   7.97e-2